+++ NEWS.md
@@ -1,7 +1,9 @@
 # eeguana 0.1.5.9000
   - Bugs: 
     - weird dplyr filter behavior  [#136](https://github.com/bnicenboim/eeguana/issues/136)
-
+    - `plot_components()` behaves more similarly to `plot_topo()`
+    - **BREAKING CHANGE** most arguments require now a `.` at the beginning: This should help to differentiate them from channels.
+    
 # eeguana 0.1.4.9000
   - Bugs: tons of compatibility issues with dplyr 1.0
   - Changes